Build better architected server
