+++ server/config/database.js
@@ -1,0 +1,7 @@
+module.exports = {
+    username: 'postgres',
+    password: 'password',
+    host: 'localhost',
+    port: 5432,
+    name: 'mattmaxwell'
+}
