+++ server/package.json
@@ -17,6 +17,7 @@
     "mongoose": "^5.9.15",
     "morgan": "~1.9.1",
     "passport": "^0.4.1",
-    "passport-jwt": "^4.0.0"
+    "passport-jwt": "^4.0.0",
+    "pg": "^8.3.0"
   }
 }
