+++ server/public/images/rotor.png
