+++ server/src/controllers/index.controller.js
@@ -1,0 +1,10 @@
+const express = require('express');
+const router = express.Router();
+
+router.get('/', (req, res, next) => {
+    console.log('Retrieving active profile...\n');
+
+    res.send('Welcome to the API!');
+});
+
+module.exports = router;
